# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FRVT Versification Viewer",
    "version": "0.1.0",
    "description": "HTTP API for the FRVT versification viewer. Every path requires HTTP Basic. This checked-in document adds a Basic security scheme and rewrites error responses to ErrorBody ({detail, code, errors?}), which is what the running server returns. GET /openapi.json on a live process is FastAPI's unpatched generator output (HTTPValidationError on many 422s, no Basic scheme). Human guide: docs/api.md."
  },
  "paths": {
...
    "/api/versifications": {
      "get": {
        "tags": [
          "versifications"
        ],
        "summary": "List Versifications",
        "description": "List schemes, optionally filtered by the ``canonical`` flag.",
        "operationId": "list_versifications_api_versifications_get",
        "parameters": [
          {
            "name": "canonical",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Canonical"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_VersificationOut_"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed. Body is ErrorBody (code validation_failed), not HTTPValidationError.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid Basic credentials. code: unauthorized. WWW-Authenticate: Basic realm=\"FRVT\".",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "429": {
            "description": "Too many failed authentications from this client IP. code: too_many_requests.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/versifications/{scheme_id}": {
      "get": {
        "tags": [
          "versifications"
        ],
        "summary": "Get Versification",
        "description": "Return one scheme including its stored ingredient.",
        "operationId": "get_versification_api_versifications__scheme_id__get",
        "parameters": [
          {
            "name": "scheme_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Scheme Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersificationDetailOut"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed. Body is ErrorBody (code validation_failed), not HTTPValidationError.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid Basic credentials. code: unauthorized. WWW-Authenticate: Basic realm=\"FRVT\".",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "429": {
            "description": "Too many failed authentications from this client IP. code: too_many_requests.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "versifications"
        ],
        "summary": "Update Versification",
        "description": "Rename a versification scheme.",
        "operationId": "update_versification_api_versifications__scheme_id__patch",
        "parameters": [
          {
            "name": "scheme_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Scheme Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VersificationUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersificationOut"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed. Body is ErrorBody (code validation_failed), not HTTPValidationError.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid Basic credentials. code: unauthorized. WWW-Authenticate: Basic realm=\"FRVT\".",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "429": {
            "description": "Too many failed authentications from this client IP. code: too_many_requests.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "versifications"
        ],
        "summary": "Delete Versification",
        "description": "Delete a scheme when no association still references it.",
        "operationId": "delete_versification_api_versifications__scheme_id__delete",
        "parameters": [
          {
            "name": "scheme_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Scheme Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Request validation failed. Body is ErrorBody (code validation_failed), not HTTPValidationError.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid Basic credentials. code: unauthorized. WWW-Authenticate: Basic realm=\"FRVT\".",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "429": {
            "description": "Too many failed authentications from this client IP. code: too_many_requests.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Page_VersificationOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/VersificationOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "Page[VersificationOut]"
      },
...
      "VersificationDetailOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "based_on_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Based On Name"
          },
          "based_on_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Based On Id"
          },
          "canonical": {
            "type": "boolean",
            "title": "Canonical"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "associated_translation_names": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Associated Translation Names"
          },
          "ingredient": {
            "additionalProperties": true,
            "type": "object",
            "title": "Ingredient"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "based_on_name",
          "based_on_id",
          "canonical",
          "created_at",
          "updated_at",
          "ingredient"
        ],
        "title": "VersificationDetailOut",
        "description": "Scheme detail including the stored Copenhagen ingredient."
      },
      "VersificationOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "based_on_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Based On Name"
          },
          "based_on_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Based On Id"
          },
          "canonical": {
            "type": "boolean",
            "title": "Canonical"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "associated_translation_names": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Associated Translation Names"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "based_on_name",
          "based_on_id",
          "canonical",
          "created_at",
          "updated_at"
        ],
        "title": "VersificationOut",
        "description": "Scheme summary without the full ingredient payload."
      },
      "VersificationUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "type": "object",
        "title": "VersificationUpdate",
        "description": "Partial update body for renaming a scheme."
      },
      "FieldError": {
        "type": "object",
        "required": [
          "field",
          "message"
        ],
        "properties": {
          "field": {
            "type": "string",
            "description": "Field or file name, for example query.limit or custom.vrs."
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ErrorBody": {
        "type": "object",
        "required": [
          "detail",
          "code"
        ],
        "description": "Envelope returned by the running server for application and framework errors. errors is omitted when there is no per-field detail.",
        "properties": {
          "detail": {
            "type": "string"
          },
          "code": {
            "type": "string",
            "enum": [
              "bad_request",
              "unauthorized",
              "not_found",
              "conflict",
              "payload_too_large",
              "validation_failed",
              "too_many_requests",
              "internal_error",
              "database_unavailable"
            ]
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "description": "Present only when there is per-field or per-file detail."
          }
        }
      }
    },
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic",
        "description": "HTTP Basic using BASIC_AUTH_USERNAME / BASIC_AUTH_PASSWORD. Local defaults admin / Admin123! must not be used on a shared host."
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local uvicorn from the repository root (see README.md)"
    }
  ],
...
  "security": [
    {
      "basicAuth": []
    }
  ]
}